move ast types into seperate ts file
